+++ time/src/test/java/org/xbib/time/chronic/SpanTestCase.java
@@ -11,14 +11,14 @@ public class SpanTestCase extends TestCase {
 
     public void testSpanWidth() {
         Span span = new Span(Time.construct(2006, 8, 16, 0), Time.construct(2006, 8, 17, 0));
-        assertEquals(60 * 60 * 24, span.getWidth());
+        assertEquals(60L * 60L * 24L, (long)span.getWidth());
     }
 
     public void testSpanMath() {
         Span span = new Span(1, 2);
-        assertEquals(2, span.add(1).getBegin());
-        assertEquals(3, span.add(1).getEnd());
-        assertEquals(0, span.subtract(1).getBegin());
-        assertEquals(1, span.subtract(1).getEnd());
+        assertEquals(2L, (long)span.add(1).getBegin());
+        assertEquals(3L, (long)span.add(1).getEnd());
+        assertEquals(0L, (long)span.subtract(1).getBegin());
+        assertEquals(1L, (long)span.subtract(1).getEnd());
     }
 }
